refactor/GETP-178: service, types 파일 네이밍 변경
