+++ src/services/people/types.ts
