+++ src/store/slice/auth.slice.ts
@@ -1,4 +1,4 @@
-import { MemberType } from "@/services/auth/auth.types";
+import { MemberType } from "@/services/auth/types";
 
 import { PayloadAction, createSlice } from "@reduxjs/toolkit";
